refactor(gui): group input feature modules
