--- crates/openlogi-gui/src/data/mod.rs
@@ -1,4 +1,0 @@
-//! Pure data: mouse button definitions, action catalog, and per-device
-//! defaults. Kept free of GPUI types so it can be reused by the CLI later.
-
-pub mod mouse_buttons;
