+++ crates/openlogi-gui/src/features/mod.rs
@@ -2,5 +2,7 @@
 
 pub mod action_ring;
 pub mod camera;
+pub mod keyboard;
 pub mod lighting;
+pub mod mouse;
 pub mod pointer;
