+++ crates/openlogi-gui/src/features/mouse/mod.rs
@@ -1,0 +1,8 @@
+//! Mouse diagram, hotspots, labels, and binding pickers.
+
+pub mod geometry;
+mod hotspots;
+pub mod leader_lines;
+pub mod picker;
+pub(crate) mod thumbwheel;
+pub mod view;
