+++ crates/openlogi-gui/src/main.rs
@@ -33,13 +33,10 @@ mod app;
 mod app_menu;
 mod asset;
 mod components;
-mod data;
 mod diagnostics;
 mod features;
 mod i18n;
 mod ipc_client;
-mod keyboard_model;
-mod mouse_model;
 mod platform;
 mod state;
 mod theme;
