--- crates/openlogi-gui/src/mouse_model/mod.rs
@@ -1,10 +1,0 @@
-//! Centre-of-screen mouse model: silhouette, hotspot triggers, action
-//! popovers, and (in later phases) leader-lines / labels and parallax.
-//!
-//! Per UI.md phases 6 (this view), 7 (leader lines), and 8 (ambient motion).
-
-pub mod geometry;
-pub mod leader_lines;
-pub mod picker;
-pub(crate) mod thumbwheel;
-pub mod view;
